Create ansible_vault secrets backend
