+++ src/commcare_cloud/secrets/backends/ansible_vault/tests/expected-generated-variables.yml
